requirements.txt
streamlit
numpy
pandas
yfinance
tensorflow-cpu==2.13.0
scikit-learn
plotly
easyocr
Pillow
python-bidi
opencv-python-headless
torch
torchvision
tqdm
